fastapi/requirements.txt
fastapi==0.115.3
keras==2.12.0
numpy==1.23.3
Pillow==9.2.0
tensorflow==2.12.0
uvicorn==0.32.0

